6.1.17 Выносит общую логику обновления в смарт компонент
